refactor: move printer modules
