+++ src/reporter/summary-reporter/finding-printer.test.ts
@@ -1,8 +1,8 @@
 import assert from "node:assert";
 import { describe, it } from "node:test";
 import { mockWritable } from "../../../test/test-utils.ts";
-import { createPrinter } from "../../core/printer.ts";
 import type { DetailedFinding } from "../../diagnostics/finding-types.ts";
+import { createPrinter } from "../../printer/printer.ts";
 import {
 	printFileFinding,
 	printLocationFinding,
