+++ src/reporter/summary-reporter/markup-printer/index.test.ts
@@ -5,7 +5,6 @@
 import { strictEqual } from "node:assert";
 import { describe, it } from "node:test";
 import { mockWritable } from "../../../../test/test-utils.ts";
-import { createPrinter } from "../../../core/printer.ts";
 import {
 	bulletList,
 	code,
@@ -19,6 +18,7 @@ import {
 	strong,
 	text,
 } from "../../../diagnostics/markup/builders.ts";
+import { createPrinter } from "../../../printer/printer.ts";
 import { printMarkup } from "./index.ts";
 
 describe("reporter/summary-reporter/markup-printer", () => {
