+++ src/reporter/summary-reporter/markup-printer/paragraph-printer.ts
@@ -2,11 +2,11 @@
  * Paragraph printing functions for markup elements.
  */
 
-import type { Printer } from "../../../core/printer.ts";
 import type {
 	MarkupGeneralElementContent,
 	MarkupParagraphElement,
 } from "../../../diagnostics/markup/types.ts";
+import type { Printer } from "../../../printer/printer.ts";
 import type { PrintContext } from "./spacing.ts";
 import { intentionToString } from "./utils.ts";
